requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jwt==1.3.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
langchain
pydantic[dotenv]
